Prevent the log parser from ignoring output.
This was done because the compilation strategy used
in some CMake projects involves compiling an object for
each source unit, and then linking all objects to create
either a (shared) library or an executable.

In doing this, CMake by default uses temporary files.
Unfortunately, the presence of compiler flags that lead
to generation of temporary files will lead to the parser
skipping that command.

As a result of this behaviour, the generator would not collect
any compiler flags at all whatsoever, and would generate an
empty configuration.